refactor(daily-notes): extract moment-format.ts pure module
Moves MOMENT_TO_LUXON + momentToLuxonFormat out of daily-notes.ts so
config-time format validation can use the converter without pulling in
filesystem or logger dependencies. Fixes the stale comment claiming
Luxon throws on unknown tokens — it renders them literally.
