+++ src/vault-mcp/vault-operations/daily-notes.ts
@@ -3,54 +3,9 @@ import { join } from "node:path"
 import { DateTime } from "luxon"
 import { logger, type Logger } from "../../logger.js"
 import { vaultFs } from "./vault-filesystem.js"
+import { momentToLuxonFormat } from "./moment-format.js"
 import { describeError } from "../../utils/describe-error.js"
 import { isErrnoException } from "../../utils/is-errno-exception.js"
-
-// ── Moment.js → Luxon format conversion ────────────────────────
-
-/** Sorted longest-first to avoid partial replacement collisions
- *  (e.g. YYYY before YY, dddd before ddd). */
-const MOMENT_TO_LUXON: ReadonlyArray<readonly [string, string]> = [
-  ["YYYY", "yyyy"],
-  ["dddd", "cccc"],
-  ["MMMM", "MMMM"],
-  ["ddd", "ccc"],
-  ["MMM", "MMM"],
-  ["YY", "yy"],
-  ["MM", "MM"],
-  ["DD", "dd"],
-  ["HH", "HH"],
-  ["hh", "hh"],
-  ["mm", "mm"],
-  ["ss", "ss"],
-  ["A", "a"],
-]
-
-/** Matches Moment.js [literal] escape groups — e.g. [Daily Note]. */
-const MOMENT_ESCAPE_RE = /\[([^\]]*)\]/g
-
-/** Converts a Moment.js format string to Luxon format tokens.
- *  Handles [literal] escapes (Moment) → 'literal' (Luxon) and
- *  common date/time tokens. Unsupported tokens (Do, d, dd) are
- *  left as-is — Luxon will throw on unknown tokens, making the
- *  failure visible rather than producing a wrong path. */
-export const momentToLuxonFormat = (momentFormat: string): string => {
-  // First pass: convert Moment [literal] escapes to Luxon 'literal' syntax.
-  // Single quotes inside literals are doubled per Luxon's escape convention.
-  const withLiteralsConverted = momentFormat.replace(
-    MOMENT_ESCAPE_RE,
-    (_, literal: string) => {
-      const escapedContent = literal.replace(/'/g, "''")
-      return `'${escapedContent}'`
-    },
-  )
-  // Second pass: replace date/time tokens from longest to shortest
-  return MOMENT_TO_LUXON.reduce(
-    (formatString, [momentToken, luxonToken]) =>
-      formatString.replaceAll(momentToken, luxonToken),
-    withLiteralsConverted,
-  )
-}
 
 // ── Config reading ──────────────────────────────────────────────
